Health page rows for the trails build and routing bundles
Per-workflow run strips (a weekly build scrolls out of the shared 40-run
window) and freshness rows read from the jobs' own single-writer health docs.
Also: search costs in Float32 (a third less memory on whole-grid fallback
searches) and a node test of the OPFS-backed PMTiles source against the
worker-built per-fire extract.

+++ frontend/src/routing/astar.ts
@@ -59,7 +59,9 @@ export class HybridSearch {
   settled = 0;
   status: SearchStatus = 'running';
   cost = Infinity;
-  private readonly g: Float64Array;
+  // Float32: seconds to ~0.01 s precision are plenty, and it cuts the
+  // largest (whole-grid fallback) search by a third on a phone.
+  private readonly g: Float32Array;
   private readonly parent: Int32Array;
   private readonly closed: Uint8Array;
   private readonly heap = new MinHeap(4096);
@@ -71,7 +73,7 @@ export class HybridSearch {
     this.wh = w.r1 - w.r0;
     this.wc = this.ww * this.wh;
     const total = this.wc + inp.graph.n;
-    this.g = new Float64Array(total).fill(Infinity);
+    this.g = new Float32Array(total).fill(Infinity);
     this.parent = new Int32Array(total).fill(-1);
     this.closed = new Uint8Array(total);
     const s = this.cellState(inp.start.cell);
